+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the "inform-job-support" extension will be documented in this file.
 
+## [Unreleased]
+### Added
+- Support Completion job name for JBI files.
+
 ## [1.1.0] - 2022-11-13
 ### Added
 - Support Japanese for validation message.